wip: accuflux with rust

+++ src/earthkit/hydro/distance/metrics.py
@@ -42,11 +42,11 @@ def min(
             "Undirected distances not yet supported for bifurcations."
         )
     if weights is None:
-        weights = np.ones(river_network.n_nodes)
+        weights = np.ones(river_network.n_nodes, dtype=np.float64)
     else:
         # maybe check sinks are all zero or nan distance
         pass
-    field = np.empty(river_network.n_nodes)
+    field = np.empty(river_network.n_nodes, dtype=np.float64)
     field.fill(np.inf)
 
     if isinstance(points, np.ndarray):
@@ -125,9 +125,11 @@ def max(
             "Max distance both upstream and downstream is not yet implemented."
         )
 
-    weights = np.ones(river_network.n_nodes) if weights is None else weights
+    weights = (
+        np.ones(river_network.n_nodes, dtype=np.float64) if weights is None else weights
+    )
 
-    field = np.empty(river_network.n_nodes)
+    field = np.empty(river_network.n_nodes, dtype=np.float64)
     field.fill(-np.inf)
 
     if isinstance(points, np.ndarray):
